+++ interfaces/curve/Gauge.sol
@@ -3,7 +3,9 @@
 pragma solidity ^0.5.17;
 
 interface Gauge {
-    function deposit(uint) external;
-    function balanceOf(address) external view returns (uint);
-    function withdraw(uint) external;
+    function deposit(uint256) external;
+
+    function balanceOf(address) external view returns (uint256);
+
+    function withdraw(uint256) external;
 }
