+++ interfaces/weth/WETH.sol
@@ -2,7 +2,9 @@ pragma solidity ^0.5.16;
 
 interface WETH {
     function deposit() external payable;
-    function withdraw(uint wad) external;
-    event Deposit(address indexed dst, uint wad);
-    event Withdrawal(address indexed src, uint wad);
+
+    function withdraw(uint256 wad) external;
+
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 }
